NeuralNetwork/requirements.txt
torch
torchvision
matplotlib
tqdm
Pillow